Now use dist_check to check all distributed related and addressed commments
